Find given queries in FM Index

Find given queries in FM Index

fm_index_find(queries, index)

Arguments

queries
Vector of strings to find in FM index

index
Index created with fm_index_construct()

Value

A data frame with three columns. pattern_index is the index of the query pattern, library_index is the index of the matching string in the index, and position is the starting position of the match in the indexed string. All indices are 1-based.

See also

Examples

data("state")
index <- fm_index_construct(state.name, case_sensitive = FALSE)
# Find all states with "new" in their names
hits <- fm_index_find("new", index)
hits

#>   pattern_index library_index position
#> 1             1            29        1
#> 2             1            30        1
#> 3             1            31        1
#> 4             1            32        1

# Show matching strings in library
state.name[hits$library_index]

#> [1] "New Hampshire" "New Jersey"    "New Mexico"    "New York"     

hits <- fm_index_find("ar", index)
hits

#>   pattern_index library_index position
#> 1             1             8        6
#> 2             1            38       12
#> 3             1             3        1
#> 4             1             4        1
#> 5             1            33        8
#> 6             1            40        8
#> 7             1            20        2

state.name[hits$library_index]

#> [1] "Delaware"       "Pennsylvania"   "Arizona"        "Arkansas"      
#> [5] "North Carolina" "South Carolina" "Maryland"